Develop - Adding decimal points to doses

+++ app/src/main/cpp/MediTrakCore/Dose/Dose.h
@@ -13,11 +13,20 @@ struct Dose {
     bool taken;
     std::string doseTime;
     std::string timeTaken;
-    int overrideDoseAmount;
+    float overrideDoseAmount;
     std::string overrideDoseUnit;
 
     Dose();
-    Dose(long id, long medicationId, bool taken, std::string doseTime, std::string timeTaken, int overrideDose = -1, std::string overrideDoseUnit = "");
+
+    Dose(
+            long id,
+            long medicationId,
+            bool taken,
+            std::string doseTime,
+            std::string timeTaken,
+            float overrideDose = -1,
+            std::string overrideDoseUnit = ""
+    );
 };
